→ Moves: Dashboard card personalization to a viewmodel slice
